requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
billiard==3.3.0.23        # via celery
boto==2.39.0
celery==3.1.23            # via django-celery
django-annoying==0.10.1
django-celery==3.1.17
django-choices==1.4.4
django-extensions==1.7.4
django-rest-swagger==0.3.10
django-sizefield==0.9.1
django-tz-detect==0.2.7
Django==1.10.1            # via django-choices, django-rest-swagger, django-sizefield, django-tz-detect
djangorestframework-jwt==1.8.0
djangorestframework==3.4.6
dogpile.cache==0.5.7      # via sxclient
dogpile.core==0.4.1       # via dogpile.cache
drf-nested-routers==0.11.1
enum34==1.1.6
futures==3.0.5
kombu==3.0.35             # via celery
py-bcrypt==0.4            # via sxclient
PyJWT==1.4.2              # via djangorestframework-jwt
python-dateutil==2.5.3
pytz==2016.6.1            # via celery, django-tz-detect
PyYAML==3.12              # via django-rest-swagger
redis==2.10.5
requests==2.9.1           # via sxclient
s3import==0.6.4
six==1.10.0               # via django-choices, django-extensions, django-tz-detect, python-dateutil, unicode-slugify
sxclient==0.18.1
unicode-slugify==0.1.3
Unidecode==0.4.19         # via unicode-slugify
whisper==0.9.15
